format / fix / new methods

+++ src/main/java/kg/attractor/jobsearch/service/ResumeService.java
@@ -19,5 +19,5 @@ public interface ResumeService {
 
     List<ResumeDto> getResumeByCategory(String category);
 
-    void editResume(Long id, ResumeDto resumeDto);
+    void editResume(Integer id, ResumeDto resumeDto);
 }
